+++ src/components/AddImageButton/AddImageButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { UseFormRegister, FieldValues, FieldError, FieldErrorsImpl, Merge } from 'react-hook-form';
-import { registerOptions } from '../../fakeData';
+import { registerOptions } from '../../data';
 import '../../components/Form/Form.scss';
 import './AddImageButton.scss';
